% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{get_readme}
\alias{get_readme}
\title{list or retrieve README documents at top level of bucket}
\usage{
get_readme(which = "populations", justList = FALSE, render = function(x)
  cat(x, sep = "\\n"))
}
\arguments{
\item{which}{character tag, suffix of README...}

\item{justList}{logical, if TRUE, just list the possible values of 'which'}

\item{render}{a function that will operate on readLines output}
}
\description{
list or retrieve README documents at top level of bucket
}
\examples{
get_readme(justList=TRUE)
get_readme()
}
