% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{lowcov_bam_urls}
\alias{lowcov_bam_urls}
\title{generate URLs for bam files of low coverage DNA sequence alignment, based on NA* or HG* sample ids}
\usage{
lowcov_bam_urls(ids, dates)
}
\arguments{
\item{ids}{character vector with sample identifiers}

\item{dates}{character vector with date tags for samples}
}
\description{
generate URLs for bam files of low coverage DNA sequence alignment, based on NA* or HG* sample ids
}
\examples{
lowcov_bam_urls(c("NA12878", "NA12812"), c("20121211", "20130415"))
}
